move folder to the src
