+++ apps/client/src/pages/index.tsx
