+++ apps/client/src/specs/index.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import Index from '../pages/index'
+import Index from '../pages'
 
 describe('Index', () => {
   it('should render successfully', () => {
